Fix #1 HW4
